Initial commit from fork of code.
This commit adds requesting ipinfo.io/json which returns
the City and State of the IP of the machine.

Additionally, the code has been modified to use Unsplash's
API to perform a random photo search based on the location
returned.

I've begun to strip out a lot of the threading as I will
likely use this as a one-off request instead of a constantly
running daemon. Longer term goals will be to integrate this
into a site or Lambda-like function.