Create initial Dockerfile
This commit adds a Dockerfile to allow creation of a docker image
to server the flask app from. Initial testing looks good.

I'll need to come up with a better scheme for secrets management
other than importing the settings file. ENV variables don't seem
to be an ideal solution either.