Use ip address for location in flask app
The flask app uses request.remote_addr to get the ip of the requesting
client and passes this to UnsplashedWallpaper's get_location method

If ip is a bogon (reserved range for private ips), we return None and
use a default location. We also use a default location for localhost
as well.

I've tested this works in Docker and runs locally as well.

We'll have to use a header parameter when this is placed behind a
proxy.