Add About Dialog screen
This commit adds a GtkAboutDialog and it is activated by the
"about" option in the MenuTray.

note: Calling out specifically that I needed to add a .hide() call
after the .run(), or else the GtkAboutDialog object is deleted from
the builder. Ended up finding a 7 year old StackOverflow article to
trace down the issue/solution.