Removing web related files and updating README.md
This commit removes the web folder, as this repo will focus on the
linux app version of the wallpaper switcher. Additionally the docker
file has been removed as well.

The readme file has been updated to reflect the changes to the
naming of the app.

I will attempt a web version of this in another repo